+++ lib/provider-retrieval-result-stats.js
@@ -1,7 +1,6 @@
 import * as Sentry from '@sentry/node'
 import timers from 'node:timers/promises'
 import pRetry from 'p-retry'
-import { buildProviderRetrievalResultStats } from './public-stats.js'
 import { createDagJsonCar } from './car.js'
 import { CID } from 'multiformats'
 
@@ -14,6 +13,22 @@ const withPgClient = fn => async ({ createPgClient, ...args }) => {
   } finally {
     await pgClient.end()
   }
+}
+
+export const build = committees => {
+  /** @type {Map<string, {total: number, successful: number}>} */
+  const providerRetrievalResultStats = new Map()
+  for (const c of committees) {
+    // IMPORTANT: include minority results in the calculation
+    for (const m of c.measurements) {
+      const minerId = m.minerId
+      const retrievalStats = providerRetrievalResultStats.get(minerId) ?? { total: 0, successful: 0 }
+      retrievalStats.total++
+      if (m.retrievalResult === 'OK') retrievalStats.successful++
+      providerRetrievalResultStats.set(minerId, retrievalStats)
+    }
+  }
+  return providerRetrievalResultStats
 }
 
 const publishRoundDetails = async ({ storachaClient, round }) => {
@@ -35,7 +50,7 @@ export const prepare = withPgClient(async ({ storachaClient, pgClient, round, co
     sparkEvaluateVersion,
     round.measurementBatches,
     roundDetailsCid.toString(),
-    buildProviderRetrievalResultStats(committees)
+    build(committees)
   ])
 })
