+++ lib/public-stats.js
@@ -1,6 +1,6 @@
 import * as Sentry from '@sentry/node'
 import createDebug from 'debug'
-
+import * as providerRetrievalResultStats from './provider-retrieval-result-stats.js'
 import { updatePlatformStats } from './platform-stats.js'
 import { getTaskId } from './retrieval-stats.js'
 
@@ -8,22 +8,6 @@ import { getTaskId } from './retrieval-stats.js'
 /** @import { Committee } from './committee.js' */
 
 const debug = createDebug('spark:public-stats')
-
-export const buildProviderRetrievalResultStats = committees => {
-  /** @type {Map<string, {total: number, successful: number}>} */
-  const providerRetrievalResultStats = new Map()
-  for (const c of committees) {
-    // IMPORTANT: include minority results in the calculation
-    for (const m of c.measurements) {
-      const minerId = m.minerId
-      const retrievalStats = providerRetrievalResultStats.get(minerId) ?? { total: 0, successful: 0 }
-      retrievalStats.total++
-      if (m.retrievalResult === 'OK') retrievalStats.successful++
-      providerRetrievalResultStats.set(minerId, retrievalStats)
-    }
-  }
-  return providerRetrievalResultStats
-}
 
 /**
  * @param {object} args
@@ -34,10 +18,10 @@ export const buildProviderRetrievalResultStats = committees => {
  * @param {(minerId: string, cid: string) => (string[] | undefined)} args.findDealClients
  */
 export const updatePublicStats = async ({ createPgClient, committees, honestMeasurements, allMeasurements, findDealClients }) => {
-  const providerRetrievalResultStats = buildProviderRetrievalResultStats(committees)
+  const stats = providerRetrievalResultStats.build(committees)
   const pgClient = await createPgClient()
   try {
-    for (const [minerId, retrievalResultStats] of providerRetrievalResultStats.entries()) {
+    for (const [minerId, retrievalResultStats] of stats.entries()) {
       await updateRetrievalStats(pgClient, minerId, retrievalResultStats)
     }
     await updateIndexerQueryStats(pgClient, committees)
